Enhance user authentication and post management features
- Integrated dotenv for environment variable management
- Updated error handling to include additional data in responses
- Refactored MongoDB connection to use environment variable
- Implemented user signup and login functionality with validation
- Added middleware for authentication checks on protected routes
- Updated user and post models to include necessary fields and relationships
- Added status management for users with corresponding routes and controllers
- Included new images and removed obsolete ones

+++ controllers/feed.js
@@ -1,7 +1,8 @@
 const { validationResult } = require("express-validator");
 
-const Post = require("../models/post");
 const { clearImage } = require("../utils/feedController.utils");
+const Post = require("../models/post");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
   const currentPage = req.query.page || 1;
@@ -45,19 +46,28 @@ exports.createPost = (req, res, next) => {
   const imageUrl = req.file.path;
   const title = req.body.title;
   const content = req.body.content;
+  let creator;
   const post = new Post({
     title: title,
     content: content,
     imageUrl: imageUrl,
-    creator: { name: "Tushar" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
-      console.log(result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully!",
-        post: result,
+        post: post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -115,6 +125,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -143,12 +158,74 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       clearImage(post.imageUrl);
       return Post.findByIdAndDelete(postId);
     })
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       console.log(result);
       res.status(200).json({ message: "Delete post." });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.getStatus = (req, res, next) => {
+  User.findById(req.userId)
+    .then((user) => {
+      if (!user) {
+        const error = new Error("User not found!");
+        error.statusCode = 404;
+        throw error;
+      }
+      res.status(200).json({ message: "Status fetched.", status: user.status });
+    })
+    .catch((err) => {
+      if (!err.statusCode) {
+        err.statusCode = 500;
+      }
+      next(err);
+    });
+};
+
+exports.updateStatus = (req, res, next) => {
+  const newStatus = req.body.status;
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    const error = new Error("Validation failed, entered data is incorrect");
+    error.statusCode = 422;
+    throw error;
+  }
+  User.findById(req.userId)
+    .then((user) => {
+      if (!user) {
+        const error = new Error("User not found!");
+        error.statusCode = 404;
+        throw error;
+      }
+      if (user.status !== newStatus.trim()) {
+        user.status = newStatus;
+        return user.save();
+      }
+      return Promise.resolve(user);
+    })
+    .then((result) => {
+      res.status(200).json({ message: "Status updated." });
     })
     .catch((err) => {
       if (!err.statusCode) {
